+++ src/index.js
@@ -1,11 +1,12 @@
 import { subscribe, Subscribe } from './subscribe'
-import { RenderPure, Provider } from './store'
+import { createStore, RenderPure, Provider } from './store'
 import ProviderContext, { createNamedContext, removeNamedContext, getNamedContext, resolveContext } from './context'
 import moduleContext from './moduleContext'
 import namedContext from './namedContext'
 import transformContext from './transformContext'
 
 export {
+    createStore,
     subscribe,
     Subscribe,
     RenderPure,